--- OddsHistory/CBB.json
@@ -1,7 +1,0 @@
-[
-
-][
-
-][
-
-]
