+++ Scrapers/Books/ESPN/cbb_espn.py
@@ -19,7 +19,7 @@ with open('../../../Dictionary/College/CBB.json', 'r') as file:
 def find_team_rank_name(dk_team_name):
     for team_mapping in team_mappings:
         if team_mapping["ESPNBet"] == dk_team_name:
-            return team_mapping["TeamID"]
+            return team_mapping["Team Rankings Name"]
     return "Unknown"  # Return a default value if not found
 
 
@@ -153,6 +153,8 @@ for z in range(1, int(number_of_games)+1):
     matchup = scrape(z)
     if matchup:
         all_matchups.append(matchup)
+    else: 
+        break
 
 print(f'Total matchups scraped: {len(all_matchups)}')
