+++ Scrapers/Books/ESPN/nba_espn.py
@@ -19,7 +19,7 @@ with open('../../../Dictionary/Pro/NBA.json', 'r') as file:
 def find_team_rank_name(dk_team_name):
     for team_mapping in team_mappings:
         if team_mapping["ESPNBet"] == dk_team_name:
-            return team_mapping["TeamID"]
+            return team_mapping["Team Rankings Name"]
     return "Unknown"  # Return a default value if not found
