+++ Scrapers/Books/ESPN/script.py
@@ -5,7 +5,7 @@ process1 = subprocess.Popen(["python", "cbb_espn.py"]) # Create and launch proce
 process2 = subprocess.Popen(["python", "mlb_espn.py"])
 process3 = subprocess.Popen(["python", "nba_espn.py"])
 
-print("Scraping DraftKings for: CBB, NBA, MLB")
+print("Scraping ESPN for: CBB, NBA, MLB")
 
 process1.wait() # Wait for process1 to finish (basically wait for script to finish)
 process2.wait()
