+++ Scrapers/Data/BetMGM/MLB.json
@@ -1,0 +1,54 @@
+[
+    {
+        "Away Team": "Angels",
+        "Home Team": "Orioles"
+    },
+    {
+        "Away Team": "Tigers",
+        "Home Team": "White Sox"
+    },
+    {
+        "Away Team": "Nationals",
+        "Home Team": "Reds"
+    },
+    {
+        "Away Team": "Yankees",
+        "Home Team": "Astros"
+    },
+    {
+        "Away Team": "Twins",
+        "Home Team": "Royals"
+    },
+    {
+        "Away Team": "Cardinals",
+        "Home Team": "Dodgers"
+    },
+    {
+        "Away Team": "Pirates",
+        "Home Team": "Marlins"
+    },
+    {
+        "Away Team": "Giants",
+        "Home Team": "Padres"
+    },
+    {
+        "Away Team": "Blue Jays",
+        "Home Team": "Rays"
+    },
+    {
+        "Away Team": "Cubs",
+        "Home Team": "Rangers"
+    },
+    {
+        "Away Team": "Guardians",
+        "Home Team": "Athletics"
+    },
+    {
+        "Away Team": "Rockies",
+        "Home Team": "Diamondbacks"
+    },
+    {
+        "Away Team": "Red Sox",
+        "Home Team": "Mariners"
+    }
+]
