+++ Scrapers/Data/BetMGM/NBA.json
@@ -1,0 +1,10 @@
+[
+    {
+        "Away Team": "Celtics",
+        "Home Team": "Hawks"
+    },
+    {
+        "Away Team": "Bucks",
+        "Home Team": "Pelicans"
+    }
+]
