--- Scrapers/Data/BetMGM/mgm.json
@@ -1,1 +1,0 @@
-[]
