+++ Scrapers/Data/DK/CBB.json
@@ -1,41 +1,41 @@
 [
     {
         "Away Team": "Clemson",
-        "Away Team Rank Name": "11ff4566",
+        "Away Team Rank Name": "Clemson",
         "DK Away Odds": {
-            "Spread": "+7.5",
-            "Spread Odds": "−112",
+            "Spread": "+7",
+            "Spread Odds": "−108",
             "Away ML": "+250"
         },
         "Home Team": "Arizona",
-        "Home Team Rank Name": "2b1ac347",
+        "Home Team Rank Name": "Arizona",
         "DK Home Odds": {
-            "Spread": "-7.5",
-            "Spread Odds": "-7.5",
+            "Spread": "-7",
+            "Spread Odds": "-7",
             "Home ML": "−310"
         },
         "Game": {
             "Start Time": "7:09PM",
-            "Total": "152",
-            "Over Total Odds": "−112",
-            "Under Total Odds": "−108",
+            "Total": "152.5",
+            "Over Total Odds": "−108",
+            "Under Total Odds": "−112",
             "League": "CBB"
         }
     },
     {
         "Away Team": "San Diego State",
-        "Away Team Rank Name": "b1915b99",
+        "Away Team Rank Name": "San Diego St",
         "DK Away Odds": {
-            "Spread": "+11",
-            "Spread Odds": "−105",
-            "Away ML": "+500"
+            "Spread": "+11.5",
+            "Spread Odds": "−112",
+            "Away ML": "+525"
         },
         "Home Team": "UConn",
-        "Home Team Rank Name": "722d49f",
+        "Home Team Rank Name": "Connecticut",
         "DK Home Odds": {
-            "Spread": "-11",
-            "Spread Odds": "-11",
-            "Home ML": "−700"
+            "Spread": "-11.5",
+            "Spread Odds": "-11.5",
+            "Home ML": "−750"
         },
         "Game": {
             "Start Time": "7:39PM",
@@ -47,18 +47,18 @@
     },
     {
         "Away Team": "Alabama",
-        "Away Team Rank Name": "96a8ead4",
+        "Away Team Rank Name": "Alabama",
         "DK Away Odds": {
-            "Spread": "+5",
-            "Spread Odds": "−115",
-            "Away ML": "+164"
+            "Spread": "+4.5",
+            "Spread Odds": "−112",
+            "Away ML": "+160"
         },
         "Home Team": "North Carolina",
-        "Home Team Rank Name": "c16ebc3c",
+        "Home Team Rank Name": "N Carolina",
         "DK Home Odds": {
-            "Spread": "-5",
-            "Spread Odds": "-5",
-            "Home ML": "−198"
+            "Spread": "-4.5",
+            "Spread Odds": "-4.5",
+            "Home ML": "−192"
         },
         "Game": {
             "Start Time": "9:50PM",
@@ -70,14 +70,14 @@
     },
     {
         "Away Team": "Illinois",
-        "Away Team Rank Name": "11b30666",
+        "Away Team Rank Name": "Illinois",
         "DK Away Odds": {
             "Spread": "+1.5",
             "Spread Odds": "−110",
             "Away ML": "+100"
         },
         "Home Team": "Iowa State",
-        "Home Team Rank Name": "838603b0",
+        "Home Team Rank Name": "Iowa St",
         "DK Home Odds": {
             "Spread": "-1.5",
             "Spread Odds": "-1.5",
