+++ Scrapers/Data/DK/MLB.json
@@ -1,14 +1,14 @@
 [
     {
         "Away Team": "LA Angels",
-        "Away Team Rank Name": "fe8acdeb",
+        "Away Team Rank Name": "LA Angels",
         "DK Away Odds": {
             "Spread": "+1.5",
             "Spread Odds": "−135",
             "Away ML": "+145"
         },
         "Home Team": "BAL Orioles",
-        "Home Team Rank Name": "e8bdce10",
+        "Home Team Rank Name": "Baltimore",
         "DK Home Odds": {
             "Spread": "-1.5",
             "Spread Odds": "-1.5",
@@ -24,14 +24,14 @@
     },
     {
         "Away Team": "STL Cardinals",
-        "Away Team Rank Name": "ccadb94e",
+        "Away Team Rank Name": "St. Louis",
         "DK Away Odds": {
             "Spread": "+1.5",
             "Spread Odds": "+100",
             "Away ML": "+190"
         },
         "Home Team": "LA Dodgers",
-        "Home Team Rank Name": "4e2f3aa1",
+        "Home Team Rank Name": "LA Dodgers",
         "DK Home Odds": {
             "Spread": "-1.5",
             "Spread Odds": "-1.5",
@@ -47,14 +47,14 @@
     },
     {
         "Away Team": "NY Yankees",
-        "Away Team Rank Name": "6424b504",
+        "Away Team Rank Name": "NY Yankees",
         "DK Away Odds": {
             "Spread": "+1.5",
             "Spread Odds": "−155",
             "Away ML": "+130"
         },
         "Home Team": "HOU Astros",
-        "Home Team Rank Name": "4005c299",
+        "Home Team Rank Name": "Houston",
         "DK Home Odds": {
             "Spread": "-1.5",
             "Spread Odds": "-1.5",
@@ -70,14 +70,14 @@
     },
     {
         "Away Team": "TOR Blue Jays",
-        "Away Team Rank Name": "8714ddbe",
+        "Away Team Rank Name": "Toronto",
         "DK Away Odds": {
             "Spread": "+1.5",
             "Spread Odds": "−185",
             "Away ML": "+114"
         },
         "Home Team": "TB Rays",
-        "Home Team Rank Name": "16f4bf06",
+        "Home Team Rank Name": "Tampa Bay",
         "DK Home Odds": {
             "Spread": "-1.5",
             "Spread Odds": "-1.5",
@@ -93,14 +93,14 @@
     },
     {
         "Away Team": "PIT Pirates",
-        "Away Team Rank Name": "9666114b",
+        "Away Team Rank Name": "Pittsburgh",
         "DK Away Odds": {
             "Spread": "+1.5",
             "Spread Odds": "−198",
             "Away ML": "+110"
         },
         "Home Team": "MIA Marlins",
-        "Home Team Rank Name": "8d47be6d",
+        "Home Team Rank Name": "Miami",
         "DK Home Odds": {
             "Spread": "-1.5",
             "Spread Odds": "-1.5",
@@ -116,14 +116,14 @@
     },
     {
         "Away Team": "MIN Twins",
-        "Away Team Rank Name": "f5328492",
+        "Away Team Rank Name": "Minnesota",
         "DK Away Odds": {
             "Spread": "-1.5",
             "Spread Odds": "+136",
             "Away ML": "−125"
         },
         "Home Team": "KC Royals",
-        "Home Team Rank Name": "db2ee960",
+        "Home Team Rank Name": "Kansas City",
         "DK Home Odds": {
             "Spread": "+1.5",
             "Spread Odds": "+1.5",
@@ -131,22 +131,22 @@
         },
         "Game": {
             "Start Time": "4:10PM",
-            "Total": "8",
-            "Over Total Odds": "−118",
-            "Under Total Odds": "−102",
+            "Total": "8.5",
+            "Over Total Odds": "−105",
+            "Under Total Odds": "−115",
             "League": "MLB"
         }
     },
     {
         "Away Team": "WAS Nationals",
-        "Away Team Rank Name": "dd4357ee",
+        "Away Team Rank Name": "Washington",
         "DK Away Odds": {
             "Spread": "+1.5",
             "Spread Odds": "−162",
             "Away ML": "+130"
         },
         "Home Team": "CIN Reds",
-        "Home Team Rank Name": "fdd0d44c",
+        "Home Team Rank Name": "Cincinnati",
         "DK Home Odds": {
             "Spread": "-1.5",
             "Spread Odds": "-1.5",
@@ -154,22 +154,22 @@
         },
         "Game": {
             "Start Time": "4:10PM",
-            "Total": "9",
-            "Over Total Odds": "−122",
-            "Under Total Odds": "+102",
+            "Total": "9.5",
+            "Over Total Odds": "−105",
+            "Under Total Odds": "−115",
             "League": "MLB"
         }
     },
     {
         "Away Team": "SF Giants",
-        "Away Team Rank Name": "486a5e3b",
+        "Away Team Rank Name": "SF Giants",
         "DK Away Odds": {
             "Spread": "-1.5",
             "Spread Odds": "+164",
             "Away ML": "−108"
         },
         "Home Team": "SD Padres",
-        "Home Team Rank Name": "3f4235d4",
+        "Home Team Rank Name": "San Diego",
         "DK Home Odds": {
             "Spread": "+1.5",
             "Spread Odds": "+1.5",
@@ -185,14 +185,14 @@
     },
     {
         "Away Team": "DET Tigers",
-        "Away Team Rank Name": "18a9d733",
+        "Away Team Rank Name": "Detroit",
         "DK Away Odds": {
             "Spread": "-1.5",
             "Spread Odds": "+100",
             "Away ML": "−170"
         },
         "Home Team": "CHI White Sox",
-        "Home Team Rank Name": "ea799920",
+        "Home Team Rank Name": "Chi Sox",
         "DK Home Odds": {
             "Spread": "+1.5",
             "Spread Odds": "+1.5",
@@ -208,14 +208,14 @@
     },
     {
         "Away Team": "CHI Cubs",
-        "Away Team Rank Name": "d10f1db8",
+        "Away Team Rank Name": "Chi Cubs",
         "DK Away Odds": {
             "Spread": "-1.5",
             "Spread Odds": "+164",
             "Away ML": "+100"
         },
         "Home Team": "TEX Rangers",
-        "Home Team Rank Name": "e841daaf",
+        "Home Team Rank Name": "Texas",
         "DK Home Odds": {
             "Spread": "+1.5",
             "Spread Odds": "+1.5",
@@ -231,18 +231,18 @@
     },
     {
         "Away Team": "CLE Guardians",
-        "Away Team Rank Name": "5a31e1ff",
+        "Away Team Rank Name": "Cleveland",
         "DK Away Odds": {
             "Spread": "-1.5",
-            "Spread Odds": "+110",
-            "Away ML": "−162"
+            "Spread Odds": "+114",
+            "Away ML": "−155"
         },
         "Home Team": "OAK Athletics",
-        "Home Team Rank Name": "8b214a89",
+        "Home Team Rank Name": "Oakland",
         "DK Home Odds": {
             "Spread": "+1.5",
             "Spread Odds": "+1.5",
-            "Home ML": "+136"
+            "Home ML": "+130"
         },
         "Game": {
             "Start Time": "10:07PM",
@@ -254,14 +254,14 @@
     },
     {
         "Away Team": "COL Rockies",
-        "Away Team Rank Name": "8f2512bc",
+        "Away Team Rank Name": "Colorado",
         "DK Away Odds": {
             "Spread": "+1.5",
             "Spread Odds": "−112",
             "Away ML": "+185"
         },
         "Home Team": "ARI Diamondbacks",
-        "Home Team Rank Name": "2b1ac347",
+        "Home Team Rank Name": "Arizona",
         "DK Home Odds": {
             "Spread": "-1.5",
             "Spread Odds": "-1.5",
@@ -277,14 +277,14 @@
     },
     {
         "Away Team": "BOS Red Sox",
-        "Away Team Rank Name": "8a2d1e91",
+        "Away Team Rank Name": "Boston",
         "DK Away Odds": {
             "Spread": "+1.5",
             "Spread Odds": "−155",
             "Away ML": "+136"
         },
         "Home Team": "SEA Mariners",
-        "Home Team Rank Name": "300faa7b",
+        "Home Team Rank Name": "Seattle",
         "DK Home Odds": {
             "Spread": "-1.5",
             "Spread Odds": "-1.5",
