+++ Scrapers/Data/DK/NBA.json
@@ -1,37 +1,37 @@
 [
     {
         "Away Team": "BOS Celtics",
-        "Away Team Rank Name": "8a2d1e91",
+        "Away Team Rank Name": "Boston",
         "DK Away Odds": {
-            "Spread": "-16.5",
-            "Spread Odds": "−108",
+            "Spread": "-16",
+            "Spread Odds": "−105",
             "Away ML": "−1350"
         },
         "Home Team": "ATL Hawks",
-        "Home Team Rank Name": "354ef58",
+        "Home Team Rank Name": "Atlanta",
         "DK Home Odds": {
-            "Spread": "+16.5",
-            "Spread Odds": "−112",
+            "Spread": "+16",
+            "Spread Odds": "−115",
             "Home ML": "+800"
         },
         "Game": {
             "Start Time": "7:40PM",
             "Total": "226",
-            "Over Total Odds": "−110",
-            "Under Total Odds": "−110",
+            "Over Total Odds": "−112",
+            "Under Total Odds": "−108",
             "League": "NBA"
         }
     },
     {
         "Away Team": "MIL Bucks",
-        "Away Team Rank Name": "c5227021",
+        "Away Team Rank Name": "Milwaukee",
         "DK Away Odds": {
             "Spread": "-1",
             "Spread Odds": "−108",
             "Away ML": "−115"
         },
         "Home Team": "NO Pelicans",
-        "Home Team Rank Name": "6a3773e4",
+        "Home Team Rank Name": "New Orleans",
         "DK Home Odds": {
             "Spread": "+1",
             "Spread Odds": "−112",
