+++ Scrapers/Data/ESPN/NBA.json
@@ -1,17 +1,17 @@
 [
     {
         "Away Team": "BOS Celtics",
-        "Away Team Rank Name": "8a2d1e91",
+        "Away Team Rank Name": "Boston",
         "ESPN Away Odds": {
-            "Spread": "-15.5",
-            "Spread Odds": "-115",
+            "Spread": "-16.5",
+            "Spread Odds": "-105",
             "Away ML": "-1600"
         },
         "Home Team": "ATL Hawks",
-        "Home Team Rank Name": "354ef58",
+        "Home Team Rank Name": "Atlanta",
         "ESPN Home Odds": {
-            "Spread": "+15.5",
-            "Spread Odds": "-105",
+            "Spread": "+16.5",
+            "Spread Odds": "-115",
             "Home ML": "+800"
         },
         "Game": {
@@ -24,14 +24,14 @@
     },
     {
         "Away Team": "MIL Bucks",
-        "Away Team Rank Name": "c5227021",
+        "Away Team Rank Name": "Milwaukee",
         "ESPN Away Odds": {
             "Spread": "-1.5",
             "Spread Odds": "+100",
             "Away ML": "-115"
         },
         "Home Team": "NO Pelicans",
-        "Home Team Rank Name": "6a3773e4",
+        "Home Team Rank Name": "New Orleans",
         "ESPN Home Odds": {
             "Spread": "+1.5",
             "Spread Odds": "-120",
